sdl_glimp: rewrite the original GL selection code
- Reuse already created SDL window for
  testing GL versions and profiles,
  but do not reuse window after testing.
- Detect best configuration possible.
- Try custom configuration if exists.
- If no custom configuration or it fails,
  load the best configuration possible.

Reuse the same window while testing for GL versions,
but destroy the window used for tests and recreate one
for the game itself. Creating a context from an SDL window
is probably modifying some window properties because
testing for both 16 bit and 24 bit context with the same
window results in a crash on Linux when the X11 server
has 16 bit depth and the GL context used to draw the
game is the same as the one used for testing GL versions.

Iterate all the known GL versions so
the log can be useful to know what's
supported.

Request the best configuration possible
so engine gfxinfo command tells the user
what is supported by his driver even
if the engine needs less.